% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_quantlim.R,
%   R/nonlinear_quantlim_orig.R
\name{nonlinear_quantlim}
\alias{nonlinear_quantlim}
\title{The goal of this function is to calculate the LoD/LoQ of the data provided in the data frame.
The function returns a new data frame containing the value of the LoD/LoQ}
\usage{
nonlinear_quantlim(
  datain,
  alpha = 0.05,
  Npoints = 100,
  Nbootstrap = 2000,
  num_changepoint_samples = 30,
  num_prediction_samples = 200,
  max_iter = 30
)

nonlinear_quantlim(
  datain,
  alpha = 0.05,
  Npoints = 100,
  Nbootstrap = 2000,
  num_changepoint_samples = 30,
  num_prediction_samples = 200,
  max_iter = 30
)
}
\arguments{
\item{num_changepoint_samples}{Number of bootstrap samples for the prediction
inteval for the changepoint. Large values can make calculations very expensive}

\item{num_prediction_samples}{Number of prediction samples to generate}

\item{max_iter}{Number of trials for convergence of every curvefit algorithm}
}
\description{
The goal of this function is to calculate the LoD/LoQ of the data provided in the data frame.
The function returns a new data frame containing the value of the LoD/LoQ
}
