% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_quantlim.R
\name{linear_quantlim}
\alias{linear_quantlim}
\title{The goal of this function is to calculate the LoD/LoQ of the data provided in
the data frame.
The function returns a new data frame containing the value of the LoD/LoQ}
\usage{
linear_quantlim(datain, alpha = 0.05, Npoints = 100, Nbootstrap = 500)
}
\description{
The goal of this function is to calculate the LoD/LoQ of the data provided in
the data frame.
The function returns a new data frame containing the value of the LoD/LoQ
}
