% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_quantlim.R
\name{nonlinear_quantlim}
\alias{nonlinear_quantlim}
\title{The goal of this function is to calculate the LoD/LoQ of the data provided in the data frame.
The function returns a new data frame containing the value of the LoD/LoQ}
\usage{
nonlinear_quantlim(datain, alpha = 0.05, Npoints = 100, Nbootstrap = 2000)
}
\description{
The goal of this function is to calculate the LoD/LoQ of the data provided in the data frame.
The function returns a new data frame containing the value of the LoD/LoQ
}
